+++ rs/execution_environment/src/scheduler.rs
@@ -2392,7 +2392,6 @@ fn get_instructions_limits_for_subnet_message(
             | ProvisionalTopUpCanister
             | UploadChunk
             | StoredChunks
-            | DeleteChunks
             | ClearChunkStore
             | TakeCanisterSnapshot
             | LoadCanisterSnapshot