+++ rs/tests/src/orchestrator/mod.rs
@@ -1,7 +1,5 @@
 pub mod node_assign_test;
-pub mod node_registration_test;
 pub mod rotate_ecdsa_idkg_key;
-pub mod ssh_access_to_nodes;
 pub mod subnet_recovery_app_subnet;
 pub mod subnet_recovery_nns_subnet;
 pub mod subnet_splitting;