+++ rs/pocket_ic_server/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+
+
+## 5.0.0 - 2024-07-22
+
 ### Added
 - A new subnet is created on an existing PocketIC instance if a new canister is created with a specified mainnet canister ID that does not belong to any existing subnet's canister range.
 - The argument of the endpoint `/http_gateway` takes an additional optional field `domains` specifying the domains at which the HTTP gateway is listening (default to `localhost`).
@@ -30,11 +34,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for canister HTTP outcalls: endpoint `/instances/<instance_id>/get_canister_http` to retrieve pending canister HTTP outcalls
   and endpoint `/instances/<instance_id>/mock_canister_http_response` to mock a response for a pending canister HTTP outcall,
   the server produces responses for pending canister HTTP outcalls automatically in the auto-progress mode (started by calling the endpoint `/instances/<instance_id>/auto_progress`).
-- Added new endpoint `/instance/<instance_id>/api/v3/canister/<effective_canister_id>/call` to support update calls to the new
-  synchronous ingress message endpoint.
+- New endpoint `/instance/<instance_id>/api/v3/canister/<effective_canister_id>/call` supporting a synchronous HTTP interface of the IC for update calls.
+  Note that this endpoint might non-deterministically return a response with status code 202 and empty body (in this case, the status of the call
+  must be polled at the endpoint `/instance/<instance_id>/api/v3/canister/<effective_canister_id>/read_state`).
   
 ### Fixed
 - Executing a query call on a new PocketIC instance crashed the PocketIC server.
+
+
 
 ## 4.0.0 - 2024-04-30
 
@@ -56,12 +63,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Potentially breaking: The time of every subnet advances by 1ns before every round execution to make sure the subnet time is strictly increasing in every round.
 
 
+
 ## 3.0.1 - 2024-02-14
 
 ### Fixed
 - Traps in tECDSA calls due to malformed tECDSA public key.
 - Server rejects jsons containing unimplemented variants of `SubnetStateConfig`.
 - The `inspect_message` method no longer panics when call is rejected.
+
+
 
 ## 3.0.0 - 2024-02-06
 
@@ -75,6 +85,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Canister inspect message errors when executing ingress messages are returned as canister execution results rather than request errors.
 - Subnets agree on which subnet id is the NNS subnet id. Fixes the problem where a canister installation via CMC directly would fail. 
+
 
 
 ## 2.0.1 - 2023-11-23